Template source refactor
