+++ packages/scafflater/scafflater.test.js
@@ -10,7 +10,7 @@ const Config = require("./scafflater-config/config");
 const RanTemplate = require("./scafflater-config/ran-template");
 const RanPartial = require("./scafflater-config/ran-partial");
 const Source = require("./scafflater-config/source");
-const TemplateSource = require("./template-source");
+const { TemplateSource } = require("./template-source");
 const { TemplateCache } = require("./template-cache");
 
 jest.mock("./scafflater-config/config");
