+++ packages/scafflater/template-manager/template-manager.test.js
@@ -1,6 +1,6 @@
 const { ScafflaterOptions } = require("../options");
 const { TemplateCache } = require("../template-cache");
-const TemplateSource = require("../template-source");
+const { TemplateSource } = require("../template-source");
 const TemplateManager = require("./template-manager");
 const {
   LocalTemplate,
