+++ packages/scafflater/template-source/github-client-template-source.js
@@ -1,5 +1,4 @@
 const LocalFolderTemplateSource = require("./local-folder-template-source");
-const GitUtil = require("../git-util");
 const fsUtil = require("../fs-util");
 const { ScafflaterOptions } = require("../options");
 const { LocalTemplate } = require("../scafflater-config/local-template");
